+++ src/extractor/extraction_containers.cpp
@@ -126,7 +126,6 @@ ExtractionContainers::ExtractionContainers()
  */
 void ExtractionContainers::PrepareData(ScriptingEnvironment &scripting_environment,
                                        const std::string &osrm_path,
-                                       const std::string &restrictions_file_name,
                                        const std::string &name_file_name)
 {
     storage::io::FileWriter file_out(osrm_path, storage::io::FileWriter::GenerateFingerprint);
@@ -139,7 +138,6 @@ void ExtractionContainers::PrepareData(ScriptingEnvironment &scripting_environme
     WriteEdges(file_out);
 
     PrepareRestrictions();
-    WriteConditionalRestrictions(restrictions_file_name);
     WriteCharData(name_file_name);
 }
 
@@ -633,16 +631,6 @@ void ExtractionContainers::WriteNodes(storage::io::FileWriter &file_out) const
     util::Log() << "Processed " << max_internal_node_id << " nodes";
 }
 
-void ExtractionContainers::WriteConditionalRestrictions(const std::string &path)
-{
-    std::uint64_t written_restriction_count = conditional_turn_restrictions.size();
-    storage::io::FileWriter restrictions_out_file(path,
-                                                  storage::io::FileWriter::GenerateFingerprint);
-    serialization::write(restrictions_out_file, conditional_turn_restrictions);
-    util::Log() << "number of conditional restrictions written to disk: "
-                << written_restriction_count;
-}
-
 void ExtractionContainers::PrepareRestrictions()
 {
 
@@ -837,7 +825,8 @@ void ExtractionContainers::PrepareRestrictions()
     const auto transform_into_internal_types =
         [&](const InputConditionalTurnRestriction &external_restriction) {
             // unconditional restriction
-            if (external_restriction.condition.empty())
+            if (external_restriction.condition.empty() &&
+                external_restriction.Type() == RestrictionType::NODE_RESTRICTION)
             {
                 TurnRestriction restriction;
                 restriction.is_only = external_restriction.is_only;
@@ -851,7 +840,9 @@ void ExtractionContainers::PrepareRestrictions()
                 restriction.is_only = external_restriction.is_only;
                 restriction.condition = std::move(external_restriction.condition);
                 if (transform(external_restriction, restriction))
+                {
                     conditional_turn_restrictions.push_back(std::move(restriction));
+                }
             }
         };